backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
pydantic[email]>=2.5.0
python-dotenv>=1.0.0
supabase>=2.0.0
hyperspell>=0.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
